+++ src/CodeSigning/CodeSigning/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added `Get-AzCodeSigningCertChain` cmdlet to retrieve the certificate chain for a certificate profile.
 
 ## Version 0.1.2
 * Updated signed 3rd party assembly Polly.dll to PSGallery